crud rewards e task

+++ RewardsForYou/Controllers/AmministratorController.cs
@@ -163,7 +163,7 @@ namespace RewardsForYou.Controllers
         { 
 
             Users deletedUser = null;
-
+            SearchDeleteUser users = new SearchDeleteUser();
             using (RewardsForYouEntities db = new RewardsForYouEntities())
             {
                 deletedUser = db.Users.Where(l => l.Serial == Serial && l.EMail == EMail).FirstOrDefault();
@@ -174,8 +174,9 @@ namespace RewardsForYou.Controllers
                     deletedUser.FiredDate = DateTime.Now;
                     db.SaveChanges();
                 }
+                users.Lista = db.Users.ToList();
             }
-                return SearchDeleteUser(new SearchDeleteUser());
+            return RedirectToAction("SearchDeleteUser", users);
         }
 
         public ActionResult TakeJsonUsers(string Serial, string EMail)
@@ -217,18 +218,12 @@ namespace RewardsForYou.Controllers
                 {
                     x = db.Rewards;
                 }
-
-               
-
+            
                 data.Lista = x.ToList();
 
-               
-
-
-
-
-                return View("SearchDeleteRewards", data);
+                
             }
+            return View("SearchDeleteRewards", data);
         }
 
        public ActionResult _PartialDeleteRewards(string Description)
@@ -244,16 +239,86 @@ namespace RewardsForYou.Controllers
         public ActionResult DoDeleteReward(string Description)
         {
             Rewards rewardDeleteted = null;
+            List<UsersRewards> usersRewards = null;
+            SearchDeleteReward rewards = new SearchDeleteReward();
             using (RewardsForYouEntities db = new RewardsForYouEntities())
             {
                 rewardDeleteted = db.Rewards.Where(l => l.Description == Description).FirstOrDefault();
-                if(rewardDeleteted != null)
+                usersRewards = db.UsersRewards.Where(l => l.RewardsID == rewardDeleteted.RewardsID).ToList();
+               
+                if(rewardDeleteted != null && usersRewards != null)
                 {
                     db.Rewards.Remove(rewardDeleteted);
+                    foreach (UsersRewards item in usersRewards)
+                    {
+                        db.UsersRewards.Remove(item);
+                    }
                     db.SaveChanges();
                 }
+                rewards.Lista = db.Rewards.ToList();
             }
-                return SearchDeleteUser(new SearchDeleteUser());
+                return RedirectToAction("SearchDeleteRewards",rewards);
+        }
+
+
+        public ActionResult SearchDeleteTasks(SearchDeleteTask data)
+        {
+
+            using (RewardsForYouEntities db = new RewardsForYouEntities())
+            {
+
+                IQueryable<Tasks> x = null;
+                if (data.Description != null)
+                {
+                    x = db.Tasks.Where(l => l.Description == data.Description);
+                }
+                else
+                {
+                    x = db.Tasks;
+                }
+                if(data.Type != null)
+                {
+                    x = db.Tasks.Where(l => l.Type == data.Type);
+                }
+                data.Lista = x.ToList();
+
+
+            }
+            return View("SearchDeleteTasks", data);
+        }
+
+        public ActionResult _PartialDeleteTasks(string Description, string Type)
+        {
+            Tasks task = null;
+            using (RewardsForYouEntities db = new RewardsForYouEntities())
+            {
+                task = db.Tasks.Where(l => l.Description == Description && l.Type == Type).FirstOrDefault();
+            }
+            return PartialView(task);
+        }
+
+        public ActionResult DoDeleteTask(string Description, string Type)
+        {
+            Tasks taskDeleted = null;
+            List<Missions> mission = null;
+            SearchDeleteTask task = new SearchDeleteTask();
+            using (RewardsForYouEntities db = new RewardsForYouEntities())
+            {
+                taskDeleted = db.Tasks.Where(l => l.Description == Description && l.Type == Type).FirstOrDefault();
+                mission = db.Missions.Where(l => l.TaskID == taskDeleted.TaskID).ToList();
+
+                if (taskDeleted != null && mission != null)
+                {
+                    db.Tasks.Remove(taskDeleted);
+                    foreach (Missions item in mission)
+                    {
+                        db.Missions.Remove(item);
+                    }
+                    db.SaveChanges();
+                }
+                task.Lista = db.Tasks.ToList();
+            }
+            return RedirectToAction("SearchDeleteTasks", task);
         }
 
     }
